split test_video.py to two files
